# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portfolio",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/team_member": {
            "get": {
                "tags": [
                    "Team member"
                ],
                "summary": "all team_member",
                "description": "team_member",
                "operationId": "team_member",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "surname": {
                                            "type": "string",
                                            "example": "surname"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "image"
                                        },
                                        "position": {
                                            "type": "string",
                                            "example": "position"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Team member"
                ],
                "summary": "team_member a new data",
                "description": "team_member new organization",
                "operationId": "00308961e568f87b2925244a9c6645e7",
                "requestBody": {
                    "description": "Team member",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "image",
                                    "position"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "name"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "surname"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "image"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "position"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/team_member/{team_member}": {
            "get": {
                "tags": [
                    "Team member"
                ],
                "summary": "team member",
                "description": "team_member_crud",
                "operationId": "76968944cd1a9665af4adce8cf7bc060",
                "parameters": [
                    {
                        "name": "team_member",
                        "in": "path",
                        "description": "team_member id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Team member"
                ],
                "summary": "Team_member a new data",
                "description": "Portfolio_Join_Image_File_crud",
                "operationId": "a064317d6b33b3c53ec33b6ff05c46c5",
                "parameters": [
                    {
                        "name": "team_member",
                        "in": "path",
                        "description": "Id of user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass category   credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "image",
                                    "position"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "name"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "surname"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "image"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "position"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Team member"
                ],
                "summary": "Delete existing project",
                "description": "Portfolio_Join_Image_File_crud",
                "operationId": "5a3a61f394606b4072b164365e66ddf7",
                "parameters": [
                    {
                        "name": "team_member",
                        "in": "path",
                        "description": "team_member id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}